Refactor markdown rendering into separate module
